fix: rename mcp directory to mcp_gateway to resolve namespace collision with pip-installed mcp package
