+++ demo_jarcore.py
@@ -17,7 +17,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parent))
 
-from mcp.coding_agent import jarcore
+from mcp_gateway.coding_agent import jarcore
 from rich.console import Console
 from rich.syntax import Syntax
 from rich.panel import Panel
