+++ docs/JARCORE_UPDATE.md
@@ -232,3 +232,9 @@ Total additions:
 - Complete documentation
 
 JARCORE = JARVIS Autonomous Reasoning & Coding Engine 🚀
+
+---
+
+## Update: Namespace Change
+
+**Note**: The `mcp/` directory referenced in this document was renamed to `mcp_gateway/` to resolve a namespace collision with the pip-installed `mcp` package. All file paths and imports in the codebase now use `mcp_gateway/` instead of `mcp/`. This change does not affect functionality - it only resolves the import conflict that prevented the official MCP SDK from being imported correctly.
