+++ issues/1.md
@@ -9,3 +9,7 @@ The current tool calling implementation relies on the Model Context Protocol (MC
 - [ ] Refactor `mcp/agent.py` and `mcp/client.py` to support UTCP.
 - [ ] Update configuration schemas in `mcp/client_config.json`.
 - [ ] Verify tool discovery and execution with the new protocol.
+
+---
+
+**Note**: File paths updated from `mcp/` to `mcp_gateway/` due to namespace collision resolution. The directory was renamed to avoid conflicts with the pip-installed `mcp` package. Current tasks should reference `mcp_gateway/` paths.
