+++ mcp_gateway/client_config.json
